Refactoring and moving submit method to parent component #689
Previously, only the submit button had the required logic for submitting a search. This logic is now required by the searchText component so, to avoid duplication, the function is being lifted to searchPageContainer, which will pass it down to searchBoxContainer and searchBoxContainerSide, which will both pass it to searchText and searchButton
